Change migration order to31, to32 -> to32, to31
